requirements.txt
fastapi==0.112.0
hashids==1.3.1
sqlmodel==0.0.21
uvicorn==0.30.5
aiomysql==0.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
requests==2.28.1
beautifulsoup4==4.12.3
lxml==5.2.2
cryptography==43.0.0
boto3==1.34.93
sqlalchemy>=2.0.14,<2.1.0
APScheduler==3.11.0
Jinja2==3.1.5
aiofiles==24.1.0
aiohttp==3.11.11
pysftp==0.2.9
pandas==2.2.3
slack-sdk==3.34.0
slowapi==0.1.9